mp8/requirements.txt
flask
boto3
Pillow
pytest
python-dotenv
requests
matplotlib
